requirements.txt
aio-pika==9.3.1
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiormq==6.7.7
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
colorama==0.4.6
frozenlist==1.4.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pamqp==3.2.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
typing_extensions==4.8.0
yarl==1.9.3
